// src/models/Story.ts
export type StoryType = {
  completed: boolean;
  contentStoryId: string;
  failVideo: string;
  instruction: string;
  introVideo: string;
  started: boolean;
  storyFinishVideo: string;
  storyOrderNumber: number;
  storyPlayedId: string;
  storySuccessText: string;
  successVideo: string;
  thumb: string;
  title: string;
  totalBodyMoves: number;
  totalTargetActualBodyMoves: number;
  totalTargetBodyMoves: number;
  totalTargetUserTimeInMillis: number;
  totalUserTimeSpentInMillis: number;
};